requirements.txt
flask
flask_sqlalchemy
flask_migrate
python-dotenv
pymysql
marshmallow
marshmallow-sqlalchemy
